Cleanup - move TextView to SwiftUITextView.swift
